// libs/enums/order.enum.ts
export enum OrderStatus {
	PENDING = 'PENDING',
	PROCESS = 'PROCESS',
	CONFIRM = 'CONFIRM',
	DELIVERED = 'DELIVERED',
	CANCEL = 'CANCEL',
}

export enum OrderItemStatus {
	PENDING = 'PENDING',
	PROCESS = 'PROCESS',
	CONFIRM = 'CONFIRM',
	DELIVERED = 'DELIVERED',
	CANCEL = 'CANCEL',
}

export enum OrderPaymentMethod {
	CREDIT_CARD = 'CREDIT_CARD',
	DEBIT_CARD = 'DEBIT_CARD',
	CASH = 'CASH',
	PAYPAL = 'PAYPAL',
	STRIPE = 'STRIPE',
	APPLE_PAY = 'APPLE_PAY',
	GOOGLE_PAY = 'GOOGLE_PAY',
	BANK_TRANSFER = 'BANK_TRANSFER',
}